embed lyrics pr merge, half assed fixes from me
