+++ app/src/main/java/pl/lambada/songsync/data/remote/UserSettingsController.kt
@@ -1,4 +1,22 @@
 package pl.lambada.songsync.data.remote
 
-class UserSettingsController {
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.core.booleanPreferencesKey
+import pl.lambada.songsync.util.get
+import pl.lambada.songsync.util.set
+
+class UserSettingsController(val dataStore: DataStore<Preferences>) {
+    var embedLyricsIntoFiles by mutableStateOf(dataStore.get(embedKey, false))
+        private set
+
+    fun setEmbedLyrics(to: Boolean) {
+        dataStore.set(embedKey, to)
+        embedLyricsIntoFiles = to
+    }
 }
+
+private val embedKey = booleanPreferencesKey("embed_lyrics")
