+++ app/src/main/java/pl/lambada/songsync/ui/Navigator.kt
@@ -3,15 +3,22 @@ package pl.lambada.songsync.ui
 import androidx.compose.animation.ExperimentalSharedTransitionApi
 import androidx.compose.animation.SharedTransitionLayout
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.platform.LocalContext
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
 import kotlinx.serialization.Serializable
 import pl.lambada.songsync.MainViewModel
+import pl.lambada.songsync.data.remote.UserSettingsController
 import pl.lambada.songsync.ui.screens.AboutScreen
 import pl.lambada.songsync.ui.screens.home.HomeScreen
+import pl.lambada.songsync.ui.screens.home.HomeViewModel
 import pl.lambada.songsync.ui.screens.search.SearchScreen
+import pl.lambada.songsync.ui.screens.search.SearchViewModel
+import pl.lambada.songsync.util.dataStore
 
 /**
  * Composable function for handling navigation within the app.
@@ -25,6 +32,10 @@ fun Navigator(
     navController: NavHostController,
     viewModel: MainViewModel
 ) {
+    val dataStore = LocalContext.current.dataStore
+    // its ok for this to recreate on config changes, no need to retain
+    val userSettingsController = remember { UserSettingsController(dataStore) }
+
     SharedTransitionLayout {
         NavHost(
             navController = navController,
@@ -33,6 +44,7 @@ fun Navigator(
             composable<ScreenHome> {
                 HomeScreen(
                     navController = navController,
+                    viewModel = viewModel { HomeViewModel(userSettingsController) },
                     sharedTransitionScope = this@SharedTransitionLayout,
                     animatedVisibilityScope = this,
                 )
@@ -41,6 +53,7 @@ fun Navigator(
                 val args = it.toRoute<ScreenSearch>()
                 SearchScreen(
                     id = args.id,
+                    viewModel = viewModel { SearchViewModel(userSettingsController) },
                     songName = args.songName,
                     artists = args.artists,
                     coverUri = args.coverUri,
