+++ app/src/main/java/pl/lambada/songsync/ui/screens/home/HomeViewModel.kt
@@ -33,6 +33,7 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.launch
+import pl.lambada.songsync.data.remote.UserSettingsController
 import pl.lambada.songsync.data.remote.lyrics_providers.others.AppleAPI
 import pl.lambada.songsync.data.remote.lyrics_providers.others.LRCLibAPI
 import pl.lambada.songsync.data.remote.lyrics_providers.others.NeteaseAPI
@@ -49,7 +50,7 @@ import java.net.UnknownHostException
 /**
  * ViewModel class for the main functionality of the app.
  */
-class HomeViewModel(private val savedStateHandle: SavedStateHandle) : ViewModel() {
+class HomeViewModel(val userSettingsController: UserSettingsController) : ViewModel() {
     private var cachedSongs: List<Song>? = null
     val selected = mutableStateListOf<String>()
     var allSongs by mutableStateOf<List<Song>?>(null)
@@ -90,7 +91,6 @@ class HomeViewModel(private val savedStateHandle: SavedStateHandle) : ViewModel(
     var showSearch by mutableStateOf(showingSearch)
 
     // selected provider
-    // TODO bring back saving
     var selectedProvider by mutableStateOf(Providers.SPOTIFY)
     val songsToBatchDownload = if (selected.isEmpty())
         displaySongs
@@ -190,37 +190,6 @@ class HomeViewModel(private val savedStateHandle: SavedStateHandle) : ViewModel(
 
     fun updateAllSongs(context: Context) = viewModelScope.launch(Dispatchers.IO) {
         allSongs = getAllSongs(context)
-    }
-
-    @SuppressLint("Range")
-    private fun getFileDescriptorFromPath(
-        context: Context, filePath: String, mode: String = "r"
-    ): ParcelFileDescriptor? {
-        val resolver: ContentResolver = context.contentResolver
-        val uri = MediaStore.Audio.Media.EXTERNAL_CONTENT_URI
-
-        val projection = arrayOf(MediaStore.Files.FileColumns._ID)
-        val selection = "${MediaStore.Files.FileColumns.DATA}=?"
-        val selectionArgs = arrayOf(filePath)
-
-        resolver.query(uri, projection, selection, selectionArgs, null)?.use { cursor ->
-            if (cursor.moveToFirst()) {
-                val fileId: Int =
-                    cursor.getInt(cursor.getColumnIndex(MediaStore.Files.FileColumns._ID))
-                if (fileId == -1) {
-                    return null
-                } else {
-                    val fileUri: Uri = Uri.withAppendedPath(uri, fileId.toString())
-                    try {
-                        return resolver.openFileDescriptor(fileUri, mode)
-                    } catch (e: FileNotFoundException) {
-                        Log.e("MainViewModel", "File not found: ${e.message}")
-                    }
-                }
-            }
-        }
-
-        return null
     }
 
     /**
@@ -395,6 +364,14 @@ class HomeViewModel(private val savedStateHandle: SavedStateHandle) : ViewModel(
         }
     }
 
+    /**
+     * Refreshes the access token by sending a request to the Spotify API.
+     */
+    suspend fun refreshToken() {
+        spotifyAPI.refreshToken()
+    }
+
+
     fun selectAllSongs() {
         ableToSelect
             ?.mapNotNull { it.filePath }
@@ -419,55 +396,6 @@ class HomeViewModel(private val savedStateHandle: SavedStateHandle) : ViewModel(
             stringPreferencesKey("blacklist"),
             blacklistedFolders.joinToString(",")
         )
-    }
-
-    fun embedLyricsInFile(
-        context: Context,
-        filePath: String,
-        lyrics: String,
-        securityExceptionHandler: (PendingIntent) -> Unit = {}
-    ): Boolean {
-        return try {
-            val fd = getFileDescriptorFromPath(context, filePath, mode = "w")
-                ?: throw IllegalStateException("File descriptor is null")
-
-            val fileDescriptor = fd.dup().detachFd()
-
-            val metadata = TagLib.getMetadata(fileDescriptor, false) ?: throw IllegalStateException(
-                "Metadata is null"
-            )
-
-            fd.dup().detachFd().let {
-                TagLib.savePropertyMap(
-                    it, propertyMap = metadata.propertyMap.apply {
-                        put("LYRICS", arrayOf(lyrics))
-                    }
-                )
-            }
-
-            true
-        } catch (securityException: SecurityException) {
-            handleSecurityException(securityException, securityExceptionHandler)
-            false
-        } catch (e: Exception) {
-            Log.e("MainViewModel", "Error embedding lyrics: ${e.message}")
-            false
-        }
-    }
-
-    private fun handleSecurityException(
-        securityException: SecurityException, intentPassthrough: (PendingIntent) -> Unit
-    ) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            val recoverableSecurityException =
-                securityException as? RecoverableSecurityException ?: throw RuntimeException(
-                    securityException.message, securityException
-                )
-
-            intentPassthrough(recoverableSecurityException.userAction.actionIntent)
-        } else {
-            throw RuntimeException(securityException.message, securityException)
-        }
     }
 }
